rename Scatter.html to Scatter.percent-range.html
It positions with percentages, so it gets the name the other percent components use. The snapshots move with it.
