+++ src/routes/_examples_ssr/Timeplot.svelte
@@ -3,7 +3,7 @@
 	import { utcDay } from 'd3-time';
 	import { scaleBand, scaleTime } from 'd3-scale';
 
-	import Scatter from '../../_components/Scatter.html.svelte';
+	import Scatter from '../../_components/Scatter.percent-range.html.svelte';
 	import AxisX from '../../_components/AxisX.percent-range.html.svelte';
 	import AxisY from '../../_components/AxisY.percent-range.html.svelte';
