+++ test/e2e/chart-screenshots.spec.js-snapshots/Snapshot-for-components-Scatter-percent-range-html-svelte-1-chromium-darwin.png
