+++ test/e2e/component-properties-table.spec.js-snapshots/Property-table-content-for-components-Scatter-percent-range-html-svelte-1.aria.yml
